Add a proxy implementation based on Kubernetes Ingresses!
We create an Endpoints, Service & Ingress object for each
of the pods. This is necessary, since Ingresses only take
Services as backends, and you need a manual Endpoint to
point a service to our pod. We can't use label selectors
on the service because the JupyterHub abstraction only
gives us IPs!